Module super_analyzer_core::results::handlebars_helpers

Handlebars helpers

Functions

all_code
Generates all the HTML for the given code.

all_lines
Generates a list of line numbers for all the given code.

generate_menu
Generates the menu for the source tree.

html_code
Generates HTML code for affected code in a vulnerability.

line_numbers
Generates a list of line numbers for the given vulnerability.

render_menu

report_index
Generates the report index for the given vulnerability.