Struct super_analyzer_core::results::report::handlebars::Report

pub struct Report {
    handler: Handlebars,
    package: String,
}

Handlebars report generator.

Fields
handler: Handlebars

Handlebars template structure.

package: String

Package name.

Methods

impl Report

pub fn from_path<P: AsRef<Path>, S: Into<String>>(
template_path: P,
package: S
) -> Result<Self, Error>

Creates a new handlebars report generator.

fn load_templates<P: AsRef<Path>>(template_path: P) -> Result<Handlebars, Error>

Loads templates from the given path.

fn generate_code_html_files(
&self,
config: &Config,
results: &Results
) -> Result<(), Error>

Generates the HTML files for the code.

fn generate_code_html_folder<P: AsRef<Path>>(
&self,
path: P,
config: &Config,
results: &Results
) -> Result<Vec<Value>, Error>

Generates a folder with HTML files with the source code of the application.

fn generate_code_html_for<P: AsRef<Path>, S: AsRef<str>>(
&self,
path: P,
config: &Config,
results: &Results,
cli_package_name: S
) -> Result<(), Error>

Generates an HTML file with source code for the given path.