Struct super_analyzer_core::static_analysis::code::Rule

struct Rule {
    regex: Regex,
    permissions: Box<[Permission]>,
    forward_check: Option<String>,
    max_sdk: Option<u32>,
    whitelist: Box<[Regex]>,
    label: String,
    description: String,
    criticality: Criticality,
    include_file_regex: Option<Regex>,
    exclude_file_regex: Option<Regex>,
}

Vulnerability searching rule.

Fields
regex: Regex permissions: Box<[Permission]> forward_check: Option<String> max_sdk: Option<u32> whitelist: Box<[Regex]> label: String description: String criticality: Criticality include_file_regex: Option<Regex> exclude_file_regex: Option<Regex>
Methods

impl Rule

pub fn regex(&self) -> &Regex

Gets the regex of the rule.

pub fn permissions(&self) -> Iter<Permission>

Gets the permissions required for this rule to be checked.

pub fn forward_check(&self) -> Option<&String>

Gets the potential forward check of the rule.

pub fn max_sdk(&self) -> Option<u32>

Gets the maximum SDK affected by this vulnerability.

pub fn label(&self) -> &str

Gets the label of the vulnerability.

pub fn description(&self) -> &str

Gets the description of the vulnerability.

pub fn criticality(&self) -> Criticality

Gets the criticality for the vulnerabilities found by the rule.

pub fn whitelist(&self) -> Iter<Regex>

Gets the whitelist regex list.

pub fn has_to_check(&self, filename: &str) -> bool

Returns if this rule has to be applied to the given filename