Module super_analyzer_core::utils

General utilities module.

Structs

Benchmark
Structure to store a benchmark information.

PARSER_CONFIG
XML parser configuration.

Statics

PARSER_CONFIG

Functions

get_code
Gets the code snippet near the start and end lines.

get_package_name
Gets the name of the package from the path of the .apk file.

get_string
Gets a string from the strings XML file.

print_vulnerability
Prints a vulnerability to stdout in a color depending on the criticality.

print_warning
Prints a warning to stderr in yellow.