Crate super_analyzer_core

SUPER Android Analyzer core library.

This library contains the code for analyzing Android applications. It's called by the launcher and contains the main logic of the analysis, with the configuration management, the logger initialization and some utility functions.

Modules

cli
Command line interface module.

config
Configuration module.

criticality
Criticality module.

decompilation
Decompilation module.

results
Results generation module.

static_analysis
Static analysis for manifest, certificate and code files.

utils
General utilities module.

Structs

Benchmark
Structure to store a benchmark information.

Config
Config structure.

PARSER_CONFIG
XML parser configuration.

Enums

ErrorKind
Enumeration of the different error kinds.

Constants

_DERIVE_failure_Fail_FOR_ErrorKind

_DERIVE_failure_core_fmt_Display_FOR_ErrorKind

Statics

BANNER
Logo ASCII art, used in verbose mode.

PARSER_CONFIG

Functions

analyze_package
Analyzes the given package with the given configuration.

copy_folder
Copies the contents of from to to

get_code
Gets the code snippet near the start and end lines.

get_package_name
Gets the name of the package from the path of the .apk file.

get_string
Gets a string from the strings XML file.

initialize_config
Initialize the config with the config files and command line options.

initialize_logger
Initializes the logger.

print_vulnerability
Prints a vulnerability to stdout in a color depending on the criticality.

print_warning
Prints a warning to stderr in yellow.